rebase updates

+++ sdk/include/opentelemetry/sdk/configuration/base2_exponential_bucket_histogram_aggregation_configuration.h
@@ -21,12 +21,9 @@ namespace configuration
 class Base2ExponentialBucketHistogramAggregationConfiguration : public AggregationConfiguration
 {
 public:
-  // TODO: max_scale range is [-10, 20]. Negative values require GetSignedInteger in the parser and
-  // changing the type of max_scale to an int32_t to match the
-  // Base2ExponentialHistogramAggregationConfig.
-  static constexpr std::size_t kDefaultMaxScale = 20;  // schema: minimum -10, maximum 20
-  static constexpr std::size_t kDefaultMaxSize  = 160;
-  static constexpr bool kDefaultRecordMinMax    = true;
+  static constexpr std::int32_t kDefaultMaxScale = 20;  // schema: minimum -10, maximum 20
+  static constexpr std::size_t kDefaultMaxSize   = 160;
+  static constexpr bool kDefaultRecordMinMax     = true;
 
   void Accept(AggregationConfigurationVisitor *visitor) const override
   {
